+++ CHANGES.md
@@ -7,6 +7,8 @@
 - Add a code action to open the closest Dune file for the current document.
   (#1817, fixes #1491, @rgrinberg)
 - Distinguish operators from functions in semantic highlighting. (#1831, @rgrinberg)
+- Support `textDocument/implementation` for type declarations in interface files.
+  (#1163)
 
 ## Fixes